# lib/api-zod/src/generated/types/account.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Accountancy Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { AccountType } from "./accountType";

export interface Account {
  id: number;
  name: string;
  type: AccountType;
  currency: string;
  balance: number;
  description?: string | null;
  createdAt: Date;
}