# lib/api-zod/src/generated/types/categoryBreakdown.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Accountancy Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryBreakdown {
  category: string;
  amount: number;
  count: number;
  percentage: number;
}