# lib/api-zod/src/generated/types/createAccountRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Accountancy Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAccountRequestType } from "./createAccountRequestType";

export interface CreateAccountRequest {
  name: string;
  type: CreateAccountRequestType;
  currency?: string;
  balance?: number;
  description?: string | null;
}