# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Accountancy Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { Transaction } from "./transaction";

export interface DashboardSummary {
  totalIncome: number;
  totalExpenses: number;
  netBalance: number;
  transactionCount: number;
  accountsCount: number;
  topCategory?: string | null;
  recentTransactions: Transaction[];
}